Refactor: move FileUtils to a separate class(Global Utils)
