+++ app/src/main/java/com/wu1015/sbnotificationbox/mailsend/MailSendActivity.java
@@ -1,14 +1,9 @@
 package com.wu1015.sbnotificationbox.mailsend;
 
-import static com.wu1015.sbnotificationbox.notification.FileUtils.delAllFiles;
-
-import android.content.Intent;
-import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Toast;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -17,10 +12,6 @@ import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
 import com.wu1015.sbnotificationbox.R;
-
-import javax.mail.Message;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeMessage;
 
 public class MailSendActivity extends AppCompatActivity {
